network, cache: Extract pod-interface to its own file
