+++ pkg/network/cache/BUILD.bazel
@@ -5,6 +5,7 @@ go_library(
     srcs = [
         "cache.go",
         "infocache.go",
+        "podinterface.go",
         "types.go",
     ],
     importpath = "kubevirt.io/kubevirt/pkg/network/cache",
@@ -25,6 +26,7 @@ go_test(
         "cache_suite_test.go",
         "cache_test.go",
         "infocache_test.go",
+        "podinterface_test.go",
         "types_test.go",
     ],
     embed = [":go_default_library"],
