+++ pkg/network/cache/infocache.go
@@ -27,11 +27,8 @@ import (
 	dutils "kubevirt.io/kubevirt/pkg/ephemeral-disk-utils"
 
 	"kubevirt.io/kubevirt/pkg/os/fs"
-	"kubevirt.io/kubevirt/pkg/util"
 	"kubevirt.io/kubevirt/pkg/virt-launcher/virtwrap/api"
 )
-
-const podIfaceCacheDirName = "network-info-cache"
 
 var virtLauncherCachedPattern = "/proc/%s/root/var/run/kubevirt-private/interface-cache-%s.json"
 var dhcpConfigCachedPattern = "/proc/%s/root/var/run/kubevirt-private/vif-cache-%s.json"
@@ -79,13 +76,6 @@ type DomainInterfaceStore interface {
 	Write(ifaceName string, cacheInterface *api.Interface) error
 }
 
-type PodInterfaceCacheStore interface {
-	IfaceEntry(ifaceName string) (PodInterfaceCacheStore, error)
-	Read() (*PodCacheInterface, error)
-	Write(cacheInterface *PodCacheInterface) error
-	Remove() error
-}
-
 type DHCPConfigStore interface {
 	Read(ifaceName string) (*DHCPConfig, error)
 	Write(ifaceName string, cacheInterface *DHCPConfig) error
@@ -106,37 +96,6 @@ func (d domainInterfaceStore) Read(ifaceName string) (*api.Interface, error) {
 func (d domainInterfaceStore) Write(ifaceName string, cacheInterface *api.Interface) (err error) {
 	err = writeToCachedFile(d.fs, cacheInterface, getInterfaceCacheFile(d.pattern, d.pid, ifaceName))
 	return
-}
-
-type PodInterfaceCache struct {
-	cache *Cache
-}
-
-func NewPodInterfaceCache(creator cacheCreator, uid string) PodInterfaceCache {
-	return PodInterfaceCache{creator.New(filepath.Join(util.VirtPrivateDir, podIfaceCacheDirName, uid))}
-}
-
-func (p PodInterfaceCache) IfaceEntry(ifaceName string) (PodInterfaceCacheStore, error) {
-	cache, err := p.cache.Entry(ifaceName)
-	if err != nil {
-		return nil, err
-	}
-
-	return PodInterfaceCache{&cache}, nil
-}
-
-func (p PodInterfaceCache) Read() (*PodCacheInterface, error) {
-	iface := &PodCacheInterface{}
-	_, err := p.cache.Read(iface)
-	return iface, err
-}
-
-func (p PodInterfaceCache) Write(cacheInterface *PodCacheInterface) error {
-	return p.cache.Write(cacheInterface)
-}
-
-func (p PodInterfaceCache) Remove() error {
-	return p.cache.Delete()
 }
 
 type dhcpConfigCacheStore struct {
