+++ pkg/network/cache/infocache_test.go
@@ -9,7 +9,6 @@ import (
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 
-	v1 "kubevirt.io/api/core/v1"
 	dutils "kubevirt.io/kubevirt/pkg/ephemeral-disk-utils"
 	"kubevirt.io/kubevirt/pkg/virt-launcher/virtwrap/api"
 )
@@ -18,7 +17,6 @@ var _ = Describe("Infocache", func() {
 
 	var tmpDir string
 	var cacheFactory *interfaceCacheFactory
-	var cacheCreator tempCacheCreator
 
 	BeforeEach(func() {
 		var err error
@@ -30,48 +28,8 @@ var _ = Describe("Infocache", func() {
 
 	AfterEach(func() {
 		os.RemoveAll(tmpDir)
-		cacheCreator.New("").Delete()
 	})
 
-	Context("PodInfoCache", func() {
-		const UID = "123"
-		var podIfaceCache PodInterfaceCacheStore
-		var cacheData PodCacheInterface
-
-		BeforeEach(func() {
-			podCache := NewPodInterfaceCache(cacheCreator, UID)
-
-			var err error
-			podIfaceCache, err = podCache.IfaceEntry("net0")
-			Expect(err).NotTo(HaveOccurred())
-
-			cacheData = PodCacheInterface{
-				Iface: &v1.Interface{
-					Model: "nice model",
-				},
-				PodIP: "random ip",
-				PodIPs: []string{
-					"ip1", "ip2",
-				},
-			}
-		})
-
-		It("should return os.ErrNotExist if no cache entry exists", func() {
-			_, err := podIfaceCache.Read()
-			Expect(err).To(MatchError(os.ErrNotExist))
-		})
-		It("should save and restore pod interface information", func() {
-			Expect(podIfaceCache.Write(&cacheData)).To(Succeed())
-			Expect(podIfaceCache.Read()).To(Equal(&cacheData))
-		})
-		It("should remove the cache file", func() {
-			Expect(podIfaceCache.Write(&cacheData)).To(Succeed())
-			Expect(podIfaceCache.Remove()).To(Succeed())
-
-			_, err := podIfaceCache.Read()
-			Expect(err).To(MatchError(os.ErrNotExist))
-		})
-	})
 	Context("DomainInfoCache", func() {
 		obj := &api.Interface{
 			Model: &api.Model{Type: "a nice model"},
